+++ src/jscc.ts
@@ -11,23 +11,27 @@ const isFunction = (fn: any): fn is Function => (!!fn && typeof fn == 'function'
 
 export default function jscc (
   code: string,
-  filename: string,
+  filename?: string | null,
   options?: JsccOptions,
   callback?: JsccCallback
 ) {
-  const props = parseOptions(filename, options)
+  // Get the normalized options
+  const props = parseOptions(filename || '', options || {})
 
-  if (isFunction(callback)) {
-    process.nextTick(() => {
-      try {
-        const result = parseBuffer(code, filename, props)
-        callback(null, result)
-      } catch (err) {
-        callback(err)
-      }
-    })
-    return undefined
+  // Run sync if not callback is given
+  if (!isFunction(callback)) {
+    return parseBuffer(code, props)
   }
 
-  return parseBuffer(code, filename, props)
+  // With a callback mimic an async behavior
+  process.nextTick(() => {
+    try {
+      const result = parseBuffer(code, props)
+      callback(null, result)
+    } catch (err) {
+      callback(err)
+    }
+  })
+  return undefined
 }
+